Lever finished (needs Testing)
Lever should be working now, if it rotates on the wrong axis or rotates to far or too little, let me know as that is an issue of just changing a few values.
More importantly makes sure normal other grabbing functions  still work as they did before as I edited the VR Pawn's grabbing functionality.